TEZ-4654: Migrate from commons-lang2.x to commons-lang3.x (#441) (Raghav Aggarwal reviewed by Laszlo Bodor)
- Upgrade to commons-lang-3.19.0
- org.apache.commons.lang.ArrayUtils => org.apache.commons.lang3.ArrayUtils
- org.apache.commons.lang.RandomStringUtils => org.apache.commons.lang3.RandomStringUtils
- org.apache.commons.lang.StringEscapeUtils => org.apache.commons.lang3.StringEscapeUtils
- org.apache.commons.lang.StringUtils => org.apache.commons.lang3.StringUtils
- org.apache.commons.lang.SystemUtils => org.apache.commons.lang3.SystemUtils
- org.apache.commons.lang.exception.ExceptionUtils => org.apache.commons.lang3.exception.ExceptionUtils
- org.apache.commons.lang.mutable.MutableInt => org.apache.commons.lang3.mutable.MutableInt
- org.apache.commons.lang.NotImplementedException => org.apache.commons.lang3.NotImplementedException

+++ tez-mapreduce/src/test/java/org/apache/hadoop/mapred/split/TestGroupedSplits.java
@@ -39,7 +39,7 @@ import java.util.Map;
 import java.util.Random;
 import java.util.Set;
 
-import org.apache.commons.lang.mutable.MutableInt;
+import org.apache.commons.lang3.mutable.MutableInt;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
@@ -915,7 +915,9 @@ public class TestGroupedSplits {
       int nodeId = Character.getNumericValue(loc.charAt(loc.length() - 1));
       assertTrue(nodeId < 4);
       assertTrue(loc.startsWith("node") && loc.length() == 5);
-      assertEquals(2, entry.getValue().getValue());
+      // commons-lang2.x MutableInt returns Object
+      // commons-lang3.x MutableInt returns Integer
+      assertEquals(2, entry.getValue().getValue().intValue());
     }
   }
